test: add find password view model unit test
* separate signup view model and find password view model
* deepen board view model test
* check all tests are passed
