+++ frontend/app/src/main/java/com/example/haengsha/model/viewModel/login/SignupViewModel.kt
@@ -1,7 +1,6 @@
 package com.example.haengsha.model.viewModel.login
 
 import androidx.lifecycle.ViewModel
-import com.example.haengsha.model.uiState.login.FindPasswordUiState
 import com.example.haengsha.model.uiState.login.SignupUiState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -49,28 +48,6 @@ class SignupViewModel : ViewModel() {
                 major = if (type == "major") newSignupData else currentState.major,
                 grade = if (type == "grade") newSignupData else currentState.grade,
                 interest = if (type == "interest") newSignupData else currentState.interest
-            )
-        }
-    }
-
-}
-
-class FindPasswordViewModel : ViewModel() {
-    private val _uiState = MutableStateFlow(FindPasswordUiState())
-    val uiState: StateFlow<FindPasswordUiState> = _uiState.asStateFlow()
-
-    fun updateEmail(email: String = "") {
-        updateFindPasswordData(email)
-    }
-
-    fun resetFindPasswordData() {
-        _uiState.value = FindPasswordUiState()
-    }
-
-    private fun updateFindPasswordData(newFindPasswordData: String) {
-        _uiState.update { currentState ->
-            currentState.copy(
-                email = newFindPasswordData
             )
         }
     }
