+++ frontend/app/src/test/java/com/example/haengsha/unitTest/login/SignupViewModelTest.kt
@@ -1,0 +1,4 @@
+package com.example.haengsha.unitTest.login
+
+class SignupViewModelTest {
+}
